float to double, student report works fine

+++ GradesServer/DTOs/ZoneScoreDto.cs
@@ -9,9 +9,9 @@
         public string ZoneName { get; set; } = string.Empty;
         public int NumNationalQuestions { get; set; }
         public int NumNationalAnsweredQuestions { get; set; }
-        public float? NationalTestScores { get; set; }
+        public double? NationalTestScores { get; set; }
         public int NumNonNationalQuestions { get; set; }
         public int NumNonNationalAnsweredQuestions { get; set; }
-        public float? NonNationalTestScores { get; set; }
+        public double? NonNationalTestScores { get; set; }
     }
 }
